requirements.txt
anthropic>=0.79.0
python-dotenv>=1.0.0
pydantic>=2.7.0
discord.py>=2.4.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiofiles>=23.2.1
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
respx>=0.21.0
